requirements.txt
groq
streamlit
chromadb
PyPDF2
#PyMuPDF
sentence-transformers
python-dotenv
#requests
#langchain-community
pypdf
